feat: uploader can now manage PDF files closes https://github.com/kalisio/kCore/issues/83 feat: media browser allow to download files closes https://github.com/kalisio/kCore/issues/84

+++ src/client/components/input/KUploader.vue
@@ -9,6 +9,7 @@
 <script>
 import _ from 'lodash'
 import 'vue2-dropzone/dist/vue2Dropzone.css'
+import 'mime-types-browser'
 import DropZone from 'vue2-dropzone'
 import { Events } from 'quasar'
 
@@ -82,7 +83,9 @@ export default {
               query: Object.assign({ resource: this.resource, resourcesService: this.resourcesService() }, this.baseQuery)
             })
             // Thumbnail as well
-            this.storageService().remove(this.files[index]._id + '.thumbnail')
+            const mimeType = mime.lookup(file.name)
+            // We only store thumbnails for images
+            if (mimeType.startsWith('image/')) this.storageService().remove(this.files[index]._id + '.thumbnail')
           }
         }
         _.pullAt(this.files, index)
@@ -90,6 +93,9 @@ export default {
       }
     },
     onThumbnailGenerated (thumbnailFile, dataUrl) {
+      const mimeType = mime.lookup(thumbnailFile.name)
+      // We only store thumbnails for images
+      if (!mimeType.startsWith('image/')) return
       const index = _.findIndex(this.files, file => file.name === thumbnailFile.name)
       if (index >= 0) {
         const id = this.generateFileId(thumbnailFile)
@@ -133,6 +139,9 @@ export default {
         formData.set(key, value)
       })
       // When not processing uploads on-the-fly send thumbnail to the server along with the file
+      const mimeType = mime.lookup(file.name)
+      // We only store thumbnails for images
+      if (!mimeType.startsWith('image/')) return
       // Check if it does exist however because it is processed asynchronously
       if (file.thumbnail) {
         this.storageService().create({ id: id + '.thumbnail', uri: file.thumbnail })
@@ -143,6 +152,11 @@ export default {
       this.addFile(_.pick(addedFile, ['name', 'size', '_id']))
       // Keep track of previews for cleanup
       this.previews.push(addedFile.previewElement)
+      const mimeType = mime.lookup(addedFile.name) 
+      if (mimeType === 'application/pdf') {
+        // This is not an image, so Dropzone doesn't create a thumbnail.
+        this.dropZoneInstance().emit('thumbnail', addedFile, this.$load('pdf-icon.png', 'asset'))
+      }
     },
     onFileUploaded (addedFile, response) {
       // We update file list on successful upload
@@ -240,16 +254,16 @@ export default {
     open (defaultFiles = []) {
       this.clear()
       // Then setup existing files on server
-      defaultFiles.forEach(file => {
+      defaultFiles.forEach(async file => {
         this.dropZoneInstance().emit('addedfile', file)
         // Make sure that there is no progress bar, etc...
         this.dropZoneInstance().emit('complete', file)
-        if (file._id) {
+        const mimeType = mime.lookup(file.name)
+        // We only generate thumbnails for images
+        if (file._id && mimeType.startsWith('image/')) {
           // Download thumbnail
-          this.storageService().get(file._id + '.thumbnail')
-          .then(image => {
-            this.dropZoneInstance().emit('thumbnail', file, image.uri)
-          })
+          const image = await this.storageService().get(file._id + '.thumbnail')
+          this.dropZoneInstance().emit('thumbnail', file, image.uri)
         }
       })
       // Because this is dynamic we need to modify the instance as the vue drop zone is not updated automatically
